refactor Labeler
Make the labeler into a regular object and use non-static methods.
Store cardinality and `MissingDataCompare` in the labeler so that it
can be set once and then passed relevant methods.

Move the label mask into its own class, and move mask creation into
the labeler so that all label logic is in one place.