fix and test masking logic
Move the masking logic into the LabelMask class, and fix it so that
it returns a label starting at bit 0, instead of in the middle.

Add tests as well. Update project settings to Java 1.7,
since the binary numbers are so nice for testing this, and will be for
testing other things as well.