remove exemplar class
This class was just a wrapper around `Instance`, and had no clear
purpose while also making the code unclear, so remove it.

Notice the strange required use of `toIntValue` in
`AnalogicalModelingOutput`. Weka uses ints in some places and doubles
in others for indexing attribute values. How annoying is that! This
may be a misunderstanding on my part, though.