fix analogical set printing and report tests
Sort the contents of the internal sets before stringification
in `AnalogicalSet` so that this output is easily testable. Fix
problems with accessing class strings, and change class/effect
maps to store the class names instead of their index doubles.

Update AnalogicalModelingOutput to print the AnalogicalSet object
instead of iterating the data and printing manually. Change it's test
to simply check that an AnalogicalSet was printed, since a test of
AnalogicalSet will probably check the details of the stringification.