fix SubcontextList tests
Update the `toString()` methods to avoid trailing commas and to produce
consistent output (via sorting). Update all tests to passing state;
`SubcontextList` was failing because its iterator was not consistent.
The fix was to use a collection and check for `contains()` returning
true, instead of explicitly following the iterator and checking results
from `hasNext()`.