discontinue use of arrays
Arrays were used in the original implementation because a fast
intersection algorithm was possible when these were naturally sorted.
The natural sorting seemed to be missing, and it would be some trouble
to figure out how to get it working. Instead of trying to squeeze more
performance out, which would probably be futile in the end anyway, go
with readability instead by replacing the use of arrays with `Set`s.

Use `HashSet`s, and add a `hashCode` method to `Subcontext` to work
with it. Remove the use of an integer index from `Subcontext`, along
with the static subcontext index that was kept there. The result is
much easier to read code. I have not profiled, but I don't care right
now.