cross-platform newline and AMconstants refactor
Store the result of `System.getProperty("line.separator")` in
AMconstants, and use it write strings in a more cross-platform-friendly
manner.

Add a utility method for formatting doubles as strings, which should
help further with cross-platform differences. Rename the class to
`AMUtils` to reflect the fact that it also contains a method.