fix lattice bug which creates extra supracontexts
There was a logic bug in the code for filling the lattice. Previously,
if the first supracontext a sub was added to became heterogeneous, it
was not added to its children. This is incorrect because the children
may have different supracontexts; say we just added to 01010 and a
previous item had the label 11010. Then the supracontext located at
11010 would be different, and it also needs to be added to so that it
is marked heterogeneous.

The correct logic is to check if the supra to be added to is ALREADY
heterogeneous, in which case we know we don't have any work at all to
do. I confirmed this by looking at the 2.1 C implementation
(amc.c, line 76).

To test that this works properly, classify a paired-down version of the
finnverb data set (the full one does not yet work correctly). Add
utilities for reducing data cardinality for testing purposes.