fix bug in timing of call to cleanSupra()
`cleanSupra()` was being called after each and every lattice insertion.
In some cases this created `Supracontext.next()` values of the current
index but which were completely unrelated to their linking
`Supracontext`, which is a violation of the algorithm's invariant.
Calling `cleanSupra()` after the insertion of an item into all of its
descendant nodes guarantees that the current index will be higher than
those of unrelated supras reachable via `getNext()`. This unifies the
implementation with that of amc.c, line 83.

Add another reduced finnverb test to show that the behavior is fixed.
This should also be added to the lattice test later.