add getSupraFromString utility test method
Creating `Supracontext` instances the long way is way too verbose,
so add a utility method that can parse the format given by
`Supracontext.toString()` and turn that into a new `Supracontext`. It
has some limitations because of limitations in Weka's `Instance` class,
but it should be much easier to use. Add a test method for it, as well,
in the same file (is there any better place to put test tests?).