refactor LatticeTest
Instead of using reflection and a weird options array, use a simple
factory to parameterize the test. Reformat the individual supracontext
tests so that they are easier to read, with the supracontext strings
next to eachother so they can be read all at once. Leave a little to-do
note, as well, since this can be cleaned-up further.