slight cleanups
Get rid of commented code that was bothering me for a long time. The
commented code made me wonder if the `Lattice` implementations were not
done, but it seems to have been just a relic of copying from the
original amc.c file. Add an `assert` in the distributed implementation
to match the basic one, and add another comment to make it clearer what
the expectation before entering that block is (and why the assert
asserts what it asserts).