reorganize label-related classes
The original labeler now extends `Labeler` as `IntLabeler`, and same
for `Label` and `IntLabel`. Public mask objects are no more; instead,
a `Labeler` object knows how to `partition` it's own labels, and that's
that. Change objects where possible to be declared instances of the
interface isntead of the implementations. Tests still needed.