band-aid the changing outcome in hetero supras
The outcome in heterogeneous supracontexts, produced by
`HeterogeneousLattice`, is not deterministic; it depends on what
subcontext it is first given, and that comes out of an unordered set.
This caused intermittent test failures. For now, the band-aid is to
specifically ignore outcomes during comparison. However, there clearly
needs to be another kind of `Supracontext` for use by
`HeterogeneousLattice`.