move hetero declaration logic into Supracontext
Move the logic for deciding if a `Supracontext` is heterogeneous or not
into the `Supracontext` class. Now the various lattices do not have to
keep track of it. Also, the outcome in heterogeneous `Supracontext`
objects will be determined by the contained data, not by the order the
data was added, so remove the workarounds from the previous commit.

This is still not a great design because now time is wasted on deciding
heterogeneity even when it is not needed because the `Supracontexts`
are not final (see `SupracontextCombiner` line 58).