fix incorrect code checking for empty supra
The code in `BasicLattice` should be checking for the known `emptySupracontext` object, not just any supra with count of zero. The algorithm will not try to decrement any other supra with count zero, and if it did it would result in an error. This unifies the code with the `DistributedLattice` implementation.