make label splitting more dynamic
Instead of always splitting into 4 or fewer lattices, aim for a maximum
cardinality of 5 per label, and split into as many partitions as are
needed to keep it there. This makes for a lattice size of 32, which
doesn't seem too bad. However, it might need tweaking in the future
(maybe 8 or 16 would be better).