fill hetero lattices in parallel
Speed up the `DistributedLattice` by filling `HeterogeneousLattice`s
in parallel using an `CompletionService`.

Update the `Labeler` implementations to be thread safe; previously
their lazy initialization of label partitioners led to threads stepping
on eachother. Also update the `Label` implementatinos to initialize
`hashCode` in the constructor instead of lazily. Lazy initialization
wasn't thread unsafe, but it did allow for multiple initializations to
happen, which is a waste of work for a method called so often.