consolidate supras in intermediate combinations
Change the `Lattice` interface to return a set instead of a list.
Update the `DistributedLattice` combining code to consolidate duplicate
supracontexts during intermediate combinations, greatly reducing the
number of intersections that must be performed. Now, the soybean set
may actually be classified, although it is still slow! Add a test for
classifying a single item from the soybean set, and note the accuracy
for the entire set in a comment instead of running the whole thing.