fix incorrect extent-adding logic
Because I did not fully understand the addIntent algorithm, I
implemented it incorrectly and it did a lot of extra work inserting
extents. Fix the algorithm to match the publication, reducing extra
work. Also change `equals` and `hashCode` in `Concept` to compare based
on intent instead of extent; using extent is incorrect because the
extent is updated after adding the concepts to parent sets, and also
concepts with different labels may possibly have the same extent.