short-circuit equals() methods
Check if `this == other` and if `other == null` in all
`equals(Object other)` methods. The first is a dirt-cheap check that
saves a deep comparison of an object with itself, and the second makes
the methods safe (and possibly faster).

This simple change shaved a second off of the runtime of the soybean
test, which is almost 20%.