Don't require git on every gradle build
The GitHub action was failing because we called `git` just to populate
the `version` variable! Read this from the `Description.props` file
instead; technically this is more correct, as this is the version that
will actually be recognized by Weka.

Also change to version format to conform to semantic versioning.