DistributedLattice hangs with fewer than 3 threads
When I designed the class, I had thought that subtasks waiting for other
subtasks would cause the waiting subtask to be paused and another run
its stead. This is incorrect! The tasks are stored in a blocking queue,
and once a task is popped from the queue it is never re-queued. Since
we have a top-level task that combines 2 sub-lattices at a time, we need
at least 3 threads available or we will block forever :((((

Honestly, I'm still not confident with this implementation. The correct
fix would be to stop using a blocking queue, and instead switch to a
mechanism that can explicitly handle sub-tasks.

Fixes #12.