Create factories for Label and Labeler
Overall this gets rid of most of the remaining reflection uses in the
test suite, and is quite a bit cleaner.

Also refactor the anonymous Labelers; one can be a static class, the
other can be a simple Mockito spy around an `IntLabeler`.

Move the lattice top and bottom retrieval logic into Labeler, since
Labeler is essentially a Label factory. I think I was going for some
kind of replacement for a generic static variable when I put it in the
label class.