Separate Lattice construction and filling steps
Move lattice filling into a separate method so that there are no checked
exceptions on the constructors. Refactor `LatticeTest` to use simple
Suppliers (which we couldn't do when there were checked exceptions to
account for!), and delete all of the extra `LatticeFactory`
implementations, as they are no longer used.

Update documentation where needed. Add a test that a Lattice does not
allow itself to be filled twice. This also exposed issues with filling
some Lattices with empty lists of subcontexts, so handle that edge case
and add a corresponding test.

This final work closes #14.