Fix thread deadlock
Calling `taskCompletionService.take().get()` from within a task running
in the same completion service can cause deadlock because the task
queue size is limited, and calling a blocking operation does not result
in a task being re-queued; this sort of thing happens at the `Executor`
level, thus my confusion, but not at the `TaskCompletionService` level,
which manages a simple queue of tasks.

Hopefully this takes care of #20. Let's see if the build passes on
all platforms 🤞.