Fix deadlock for reals
Oookay, I was wrong in 3c8ba1dd91, the queue of running tasks is
stored in the `Executor`, not in the `CompletionService`, which just
provides a nice clean API for getting the next finished result.

I learned to correct way to schedule subtasks from Baeldung:
https://www.baeldung.com/java-work-stealing. The subtasks have to be
a subclass of `ForkJoinTask`, and blocking on a subtask must be done
using the `join` method, which signals to the `ForkJoinPool` that the
current task is blocked and should be rescheduled. This is called
"managed blocking", and was surprisingly difficult in Java. Apparently
in Scala you can just use a `blocking {}` code block and it will be
managed for you.

Fixes #20.