Add context and instance printing to labeler
These will be used as utility methods for analogical set and gangs
printing.

Following Algorithm::AM's format for context labels, we use "val val
val *" for context labels, where "*" indicates that a particular
attribute value does not match the test item, and "val" indicates that
the attribute does match.

Fix sixCardinalityData, which was creating numeric attributes instead of
nominal, which is what AM uses and is required for proper context label
printing.