Store human-friendly label string in Subcontext
This was the cleanest way to use the string in generated reports, since
we wanted to sort using the string.

Always adding the human-readable label means that the method gets called
much more often, which lead to more corner cases being found. The
`toString()` method for `BitSetLabel`, for example, was utterly broken,
and none of the labels could handle empty attribute sets, which is a
case exercised by Weka's classifier test.