Fix distribution build
Rename the `dist` target and document its usage so that I don't forget
about it next time! Fix a minor JavaDoc error (JavaDoc is included in
the dist), and update the gradle target to create a fat jar. Originally
I submitted the distribution to Mark and the plugin failed to load
because the picnic dependency was not included.