override `printClassification` in `AMOutput`
One of those times when I feel really frustrated working with Weka.

`doPrintClassification(Classifier...)` is an abstract method and
therefore must be implemented in `AnalogicalModelingOutput`. However,
the `ClassifierPanel` actually calls
`printClassification(Classifier...)`, which in turn calls
`doPrintClassification(double[]...`, which is useless for reporting the
information we need to. This means that unless we override
`printClassification`, the `doPrintClassification` method that we need,
and which is required by the Weka API to be implemented, is dead code.
:(((

The fix is simple, but figuring it out was quite difficult (glad that
Mark is so willing to help).

Override `printClassification` so that it calls the correct
`doPrintClassification` method. Throw an exception in the useless-but-
required version of `doPrintClassification` version so that we'll know
immediately next time this issue occurrs (assuming, of course, that Weka
will properly report exception messages...).