Honor the numDecimals setting in AMOutput
`AbstractOutput` has a setting `numDecimals` for specifying the
precision of floating point values in the output. Refactor so that this
setting can be honored everywhere.

Java does not have infinite-precision BigDecimal division, so until the
printed precision is known we use `DECIMAL64`, which is the best most
machines can do.

Using DECIMAL64 instead of DECIMAL32 caused a few tests in
`AMResultsTest` to fail; this is easily fixed by using the EPSILON value
of 1e-7, which correctly matches the number of decimal places that are
being specified in the asserted values specifications.