Update readme
Thanks to Deborah Roughton for trying out AM and sharing what she found
easy or difficult to use:
https://www.youtube.com/watch?v=RQS57r5CqDw&ab_channel=DeborahRoughton

I'm glad to know that the instructions in the Readme were very user
friendly, but we needed the following information:

1) The types of attributes that AM can handle and how to pre-process a
dataset so that AM can handle
2) AnalogicalModelingOutput can only be used with AnalogicalModeling

Since we've also updated the output recently, also add
information on configuring it as well as a screenshot showing the
analogical sets and gang effects in the output.

Also fix typos and formatting, add a pointer to the documentation on
GitHub pages, mention the CI builds, and update the gradle build info.