Fix `distributionForInstance` indexing
The numbers returned in `distributionForInstance` were essentially
ordered randomly, though ordering for the simple chapter 3 data happened
to always be correct. I probably didn't realize when I wrote the method
that the ordering mattered; or maybe this was something I meant to fix
later? Either way, it was completely non-sensical, relying on the
ordering of iteration over a `HashMap`.

Besides fixing this bug, update the unit tests so that they test the
return value, rather than just the stored `AMResults`, which completely
separate from the reporting done by weka.