Handle ignoreUnknowns in classifier options
Previously, if you selected "ignore unknowns" in the GUI, it would not
be reflected in the option parameters.

Add a new option, `-I`, to set the `ignoreUnknowns` parameter. Fix the
tooltip and the parameter name to the `get` method, as well.

Add simple tests for the generated option array.

Fixes #34.