implement CSV formatting for distribution report
2 new concepts were required for the reporting that Sabine requires:

* Number of exemplars in the training set was never tracked before. We
now track this in the subcontext list, which we now store in the
AMResults class.
* Judgement: correct, tie, incorrect, etc. I used this term because I
found "prediction" to be easily confusable with the predicted class.

There is also some work on a CSV document header, but the distribution
is only a one-row document, so the header would be overkill. I can
always add it in the follow-up Python tool for collecting all of these
CSV documents in one place.